requirements.txt
appdirs==1.4.4
attrs==19.3.0
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
idna==2.9
multidict==4.7.5
pathspec==0.8.0
python-dateutil==2.8.1
regex==2020.5.7
requests==2.23.0
six==1.14.0
soupsieve==2.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.9(yarl==1.4.2
